Document and further implement oop in configurations
